[WIP] Move field base classes and add docstrings
